Factor out common code for test storage and optimization checking.
Reuse in `B2Dataset` tests too.
